pyproject.toml
[tool.poetry]
name = "azure-ai-services-solutions"
version = "0.1.0"
description = "A GitHub template repository for a Python"
authors = ["ks6088ts <ks6088ts@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
ruff = "^0.3.5"
black = "^24.4.0"
pytest = "^8.1.1"
coverage = "^7.4.4"
pytest-cov = "^5.0.0"


[tool.poetry.group.backend.dependencies]
fastapi = "^0.110.3"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
openai = "^1.25.0"
azure-identity = "^1.16.0"
pydantic-settings = "^2.2.1"
azure-ai-documentintelligence = "^1.0.0b3"
python-multipart = "^0.0.9"
azure-storage-blob = "^12.19.1"
azure-ai-vision-imageanalysis = "^1.0.0b2"
azure-eventgrid = "^4.19.0"


[tool.poetry.group.frontend.dependencies]
streamlit = "^1.33.0"
aiohttp = {extras = ["speedups"], version = "^3.9.5"}
azure-cognitiveservices-speech = "^1.37.0"
microsoft-kiota-abstractions = "^1.3.2"
microsoft-kiota-authentication-azure = "^1.0.0"
microsoft-kiota-http = "^1.3.1"
microsoft-kiota-serialization-form = "^0.1.0"
microsoft-kiota-serialization-json = "^1.2.0"
microsoft-kiota-serialization-multipart = "^0.1.0"
microsoft-kiota-serialization-text = "^1.0.0"


[tool.poetry.group.azure-functions.dependencies]
azure-functions = "^1.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = ["client"]
respect-gitignore = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["D203"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
skip_glob = ["client/*"]

[tool.pytest.ini_options]
addopts = "-ra --cov"
testpaths = ["tests"]
pythonpath = ['.']

[tool.coverage.report]
show_missing = true
